complete app*

+++ app/src/main/java/com/example/runningtracker/model/User.kt
@@ -2,5 +2,5 @@ package com.example.runningtracker.model
 
 import java.io.Serializable
 
-data class User(val name: String, val weight: Int): Serializable {
+data class User(val name: String, val weight: Double): Serializable {
 }
